Added default config method for nodes via json.
Default configuration method uses GLib's type system to compare types in
the json nodes and the apollo Nodes.  All types that can be introspected
with a ParamSpec must pass for the default config to determine that the
configuration was successful.  The default config method can still be
overridden, but this should work for all non-array based methods.