Added logic for dynamic tree generation by name.
Gave BehavioralTreeSet the ability to instantiate nodes based on their
name as registered by the GLib type system.  All nodes now are required
to implement the 'void init()' function in place of the constructor.
This should help to clean up the code necessary in the Node.configure()
at a later point in time.  A test is added to check that the tree and
the node generation is correct.